% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsApi.R
\name{api_rfit}
\alias{api_rfit}
\title{Api connect to api.regression.fit to fetch datasets
ELo 202205 GWU DATS
version 1.0}
\usage{
api_rfit(table)
}
\arguments{
\item{table}{The name of table on DB}
}
\value{
dataframe
}
\description{
Api connect to api.regression.fit to fetch datasets
ELo 202205 GWU DATS
version 1.0
}
\examples{
api_rfit( 'gapminder' )
table names can be in this list ('AStudentRecord','BikeShare','Dats_grades','Diet6wk','Happy','gapminder','gradAdmit','nfl2008_fga','Pizza','Titanic','USDANutrient','GSS_demographics','AAPL_daily','AAPL_full','GOOG_full','GOOG_daily','MSFT_daily','MSFT_full','GSS_demographics_xlsx','World1','World2');
updated list can also be found at https://api.regression.fit/endpt.json
}
