% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsLoader.R
\name{loadPkg}
\alias{loadPkg}
\title{use this function to conveniently load libraries and work smoothly with knitting
can add quietly=T option to the require() function
the loadPkg function essentially replaced/substituted two functions install.packages() and library() in one step.
alternative CRAN example, "http://cran.rstudio.com"
Combining install.packages() (if needed) with library() in one, knitr friendly.
ELo 2019 GWU DATS}
\usage{
loadPkg(pkg, reposurl = "http://cran.us.r-project.org")
}
\arguments{
\item{pkg}{Name of package}

\item{reposurl}{CRAN/repo url}
}
\description{
use this function to conveniently load libraries and work smoothly with knitting
can add quietly=T option to the require() function
the loadPkg function essentially replaced/substituted two functions install.packages() and library() in one step.
alternative CRAN example, "http://cran.rstudio.com"
Combining install.packages() (if needed) with library() in one, knitr friendly.
ELo 2019 GWU DATS
}
\examples{
loadPkg("ggplot2", "http://cran.rstudio.com")
}
