% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsOutliers.R
\name{outlierKD}
\alias{outlierKD}
\title{outlierKD function by By Klodian Dhana,
https://www.r-bloggers.com/identify-describe-plot-and-remove-the-outliers-from-the-dataset/}
\usage{
outlierKD(dt, var)
}
\description{
outlierKD function by By Klodian Dhana,
https://www.r-bloggers.com/identify-describe-plot-and-remove-the-outliers-from-the-dataset/
}
