% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsOutliers.R
\name{outlierKD2}
\alias{outlierKD2}
\title{Original outlierKD function by By Klodian Dhana,
https://www.r-bloggers.com/identify-describe-plot-and-remove-the-outliers-from-the-dataset/
Modified to have third argument for removing outliers instead of interactive prompt,
and after removing outlier, original df will not be changed. The function returns the a df,
which can be saved as original df name if desired.
Also added QQ-plot in the output, with options to show/hide boxplot, histogram, qqplot.
Check outliers, and option to remove them, save as a new dataframe.}
\usage{
outlierKD2(
  df,
  var,
  rm = FALSE,
  boxplt = FALSE,
  histogram = TRUE,
  qqplt = FALSE
)
}
\arguments{
\item{df}{The dataframe.}

\item{var}{The variable in the dataframe to be checked for outliers}

\item{rm}{Boolean. Whether to remove outliers or not.}

\item{boxplt}{Boolean. Whether to show the boxplot, before and after outliers removed.}

\item{histogram}{Boolean. Whether to show the histogram, before and after outliers removed.}

\item{qqplt}{Boolean. Whether to show the qqplot, before and after outliers removed.}
}
\value{
The dataframe with outliers replaced by NA if rm==TRUE, or df if nothing changed
}
\description{
Original outlierKD function by By Klodian Dhana,
https://www.r-bloggers.com/identify-describe-plot-and-remove-the-outliers-from-the-dataset/
Modified to have third argument for removing outliers instead of interactive prompt,
and after removing outlier, original df will not be changed. The function returns the a df,
which can be saved as original df name if desired.
Also added QQ-plot in the output, with options to show/hide boxplot, histogram, qqplot.
Check outliers, and option to remove them, save as a new dataframe.
}
\examples{
  outlierKD2(mydf, height, FALSE, TRUE, TRUE, TRUE)
  mydf = outlierKD2(mydf, height, TRUE, TRUE, TRUE, TRUE)
  mydfnew = outlierKD2(mydf, height, TRUE)
}
