% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsXform.R
\name{PCAxform}
\alias{PCAxform}
\title{PCA transform function
Obtain the dataframe with the Principal Components after the rotation.
ELo 201911 GWU DATS}
\usage{
PCAxform(df, z = TRUE)
}
\arguments{
\item{df}{The dataframe.}

\item{z}{T/F or 0/1 for z-score to be used}
}
\value{
The transformed dataframe.
}
\description{
PCA transform function
Obtain the dataframe with the Principal Components after the rotation.
ELo 201911 GWU DATS
}
\examples{
tmp = PCAxform(USArrests,TRUE)
}
