% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsXform.R
\name{PCRxform}
\alias{PCRxform}
\title{Obtain the dataframe with the Principal Components after the rotation for PCRegression. Requires related function PCAxform()
ELo 201903 GWU DATS}
\usage{
PCRxform(df, y, zX = TRUE, zy = FALSE)
}
\arguments{
\item{df}{The dataframe.}

\item{y}{The y-variable column index number(int), or the name of y-variable}

\item{zX}{T/F or 0/1 for z-score used on X-variables}

\item{zy}{T/F or 0/1 for z-score used on the target y-variable}
}
\value{
The transformed dataframe.
}
\description{
Obtain the dataframe with the Principal Components after the rotation for PCRegression. Requires related function PCAxform()
ELo 201903 GWU DATS
}
\examples{
tmp = PCAxform(USArrests,TRUE)
}
