% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsLoader.R
\name{unloadPkg}
\alias{unloadPkg}
\title{unload packages when done
ELo 2019 GWU DATS}
\usage{
unloadPkg(pkg, character.only = FALSE)
}
\arguments{
\item{pkg}{Name of package}

\item{character.only}{character.only = FALSE default}
}
\description{
unload packages when done
ELo 2019 GWU DATS
}
\examples{
unloadPkg("ggplot2")
}
