% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsXform.R
\name{uzscale}
\alias{uzscale}
\title{Universal z-scale function
Standardize dataframe to z scores, safe for non-numeric variables.
ELo 201904 GWU DATS}
\usage{
uzscale(df, append = 0, excl = NULL)
}
\arguments{
\item{df}{The dataframe.}

\item{append}{T/F or 0/1. Option to append scaled columns or replace original columns in the dataframe.}

\item{excl}{A list c(a,b,"d","ef") of excluded columns, either by their indexes and/or names.}
}
\value{
The transformed dataframe, appended or replaced with standardized scores. Non-numeric columns will not be appended, or if "replace option" is chosen, the columns will be untouched.
}
\description{
Universal z-scale function
Standardize dataframe to z scores, safe for non-numeric variables.
ELo 201904 GWU DATS
}
\examples{
  tmp = uzscale( ISLR::Hitters )
  tmp = uzscale( ISLR::Hitters, 1 )
  tmp = uzscale( ISLR::Hitters, TRUE, c(19,"NewLeague") )
}
