% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsDisplay.R
\name{xkabledply}
\alias{xkabledply}
\title{Combining base::summary, xtable, and kableExtra, to easily display model summary.
wrapper for the base::summary function on model objects
Can also use as head/tail for nicer display
ELo 202004 GWU DATS
version 1.2.2}
\usage{
xkabledply(
  modelsmmrytable,
  title = "",
  digits = 4,
  pos = "left",
  bso = "striped",
  wide = FALSE
)
}
\arguments{
\item{modelsmmrytable}{This can be a generic table, a model object such as lm(), or the summary of a model object summary(lm())}

\item{title}{Title of table. If modelsmmrytable has a "formula", it will be the default title}

\item{digits}{Number of digits to display}

\item{pos}{Position of table, c("left","center","right")}

\item{bso}{bootstrap_options = c("basic", "striped", "bordered", "hover", "condensed", "responsive")}

\item{wide}{print table in long (FALSE) format or wide (TRUE) format}
}
\value{
HTML table for display
}
\description{
Combining base::summary, xtable, and kableExtra, to easily display model summary.
wrapper for the base::summary function on model objects
Can also use as head/tail for nicer display
ELo 202004 GWU DATS
version 1.2.2
}
\examples{
xkabledply( df, title="Table testing", pos="left", bso="hover" )
xkabledply( ISLR::Hitters[1:5,] )
}
