% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsDisplay.R
\name{xkabledplyhead}
\alias{xkabledplyhead}
\title{Better display than the default head() function
ELo 202004 GWU DATS}
\usage{
xkabledplyhead(
  df,
  rows = 5,
  title = "Head",
  digits = 4,
  pos = "left",
  bso = "striped"
)
}
\arguments{
\item{df}{Dataframe}

\item{title}{Title of table.}

\item{digits}{Number of digits to display}

\item{pos}{Position of table, c("left","center","right")}

\item{bso}{bootstrap_options = c("basic", "striped", "bordered", "hover", "condensed", "responsive")}
}
\value{
HTML table for display
}
\description{
Better display than the default head() function
ELo 202004 GWU DATS
}
\examples{
xkabledplyhead( ISLR::Hitters[1:5,], title="Head of df", pos="left", bso="hover" )
}
