% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsDisplay.R
\name{xkablesummary}
\alias{xkablesummary}
\title{Combining base::summary, xtable, and kableExtra, to easily display numeric variable summary of dataframes.
ELo 202004 GWU DATS
version 1.2}
\usage{
xkablesummary(
  df,
  title = "Table: Statistics summary.",
  digits = 4,
  pos = "left",
  bso = "striped"
)
}
\arguments{
\item{df}{The dataframe.}

\item{title}{Title of table.}

\item{digits}{Number of digits to display}

\item{pos}{Position of table, c("left","center","right")}

\item{bso}{bootstrap_options = c("basic", "striped", "bordered", "hover", "condensed", "responsive")}
}
\value{
The HTML summary table for display, or for knitr to process into other formats
}
\description{
Combining base::summary, xtable, and kableExtra, to easily display numeric variable summary of dataframes.
ELo 202004 GWU DATS
version 1.2
}
\examples{
  xkablesummary( ISLR::Hitters, title="Five number summary", pos="left", bso="hover"  )
}
