% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsDisplay.R
\name{xkablevif}
\alias{xkablevif}
\title{Combining faraway::vif, xtable, and kableExtra, to easily display numeric summary of VIFs for a model.
ELo 202004 GWU DATS
version 1.3}
\usage{
xkablevif(
  model,
  title = "",
  digits = 3,
  pos = "left",
  bso = "striped",
  wide = TRUE
)
}
\arguments{
\item{model}{The lm or compatible model object.}

\item{title}{Title of table.}

\item{digits}{Number of digits to display}

\item{pos}{Position of table, c("left","center","right")}

\item{bso}{bootstrap_options = c("basic", "striped", "bordered", "hover", "condensed", "responsive")}

\item{wide}{print table in long (FALSE) format or wide (TRUE) format}
}
\value{
The HTML summary table of the VIFs for a model for display, or for knitr to process into other formats
}
\description{
Combining faraway::vif, xtable, and kableExtra, to easily display numeric summary of VIFs for a model.
ELo 202004 GWU DATS
version 1.3
}
\examples{
xkablevif( lm(Salary~Hits+RBI, data=ISLR::Hitters), wide=T )
}
